INFUND-95 added the javax-persistency dependency for the api lib

+++ ifs-data-service/src/main/java/com/worth/ifs/repository/UserRepository.java
@@ -14,7 +14,7 @@ import java.util.List;
 public interface UserRepository extends PagingAndSortingRepository<User, Long> {
 
         List<User> findByName(@Param("name") String name);
-        List<User> findById(@Param("id") Integer id);
+        List<User> findById(@Param("id") Long id);
         List<User> findByToken(@Param("token") String token);
